Metodi del Calcolo Scientifico/Progetto 2/requirements.txt
openpyxl
numpy
scipy
matplotlib
pandas
pysimplegui
pillow